requirements.txt
flask==2.3.3
flask-cors==4.0.0
azure-cosmosdb-table==1.0.6
azure-functions==1.15.0
mutagen==1.47.0
requests==2.31.0